+++ packages/pwa-kit-create-app/CHANGELOG.md
@@ -1,5 +1,6 @@
-## v3.16.0-dev (Dec 17, 2025)
-- Update `default.js` and `/_app-config/index.jsx` template to use email mode by default for passwordless login. [#3526] (https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3526)
+
+## v3.15.0-dev (Nov 05, 2025)
+- Update `default.js` and `/_app-config/index.jsx` template to use email mode by default for passwordless login and password reset. [#3526] (https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3526) [#3547](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3547)
 
 ## v3.15.0 (Dec 17, 2025)
 - Add new Google Cloud API configuration and Bonus Product configuration [#3523](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/3523)